% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factory.R
\name{as_iqr_theme}
\alias{as_iqr_theme}
\title{Convert theme specification to a ggplot2 theme object}
\usage{
as_iqr_theme(theme = NULL)
}
\arguments{
\item{theme}{A character string (theme name), a function (returning a ggplot2 theme),
an IqrTheme object, or NULL.}
}
\value{
A ggplot2 theme object (can be directly added with \code{+}).
}
\description{
Convert theme specification to a ggplot2 theme object
}
