% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factory.R
\name{combine_plots}
\alias{combine_plots}
\title{Combine Multiple Plots into a Dashboard Layout}
\usage{
combine_plots(..., plotlist = NULL, ncol = NULL, nrow = NULL, theme = NULL)
}
\arguments{
\item{...}{plot objects}

\item{plotlist}{list of plot objects}

\item{ncol}{number of columns}

\item{nrow}{number of rows}

\item{theme}{theme specification, applied to all plots}
}
\value{
combined plot (patchwork)
}
\description{
Combine Multiple Plots into a Dashboard Layout
}
