% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pareto_ok.R
\name{create_pareto_plot}
\alias{create_pareto_plot}
\title{Create Pareto Plot}
\usage{
create_pareto_plot(
  data,
  category_col = NULL,
  count_col = NULL,
  group_col = NULL,
  weight_col = NULL,
  iqr_theme,
  title = NULL,
  subtitle = NULL,
  ...
)
}
\arguments{
\item{data}{Prepared data.frame}

\item{category_col}{Original category column name}

\item{count_col}{Original count column name}

\item{group_col}{Original group column name}

\item{weight_col}{Original weight column name}

\item{iqr_theme}{IqrTheme object}

\item{title}{Chart title}

\item{subtitle}{Chart subtitle}

\item{...}{Additional arguments}
}
\value{
ggplot object
}
\description{
Internal function to create the Pareto chart (top part).
}
