% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pareto_ok.R
\name{create_pareto_table}
\alias{create_pareto_table}
\title{Create Pareto Table}
\usage{
create_pareto_table(
  data,
  category_col = NULL,
  count_col = NULL,
  group_col = NULL,
  weight_col = NULL,
  iqr_theme
)
}
\arguments{
\item{data}{Prepared data.frame}

\item{category_col}{Original category column name}

\item{count_col}{Original count column name}

\item{group_col}{Original group column name}

\item{weight_col}{Original weight column name}

\item{iqr_theme}{IqrTheme object}
}
\value{
ggplot object
}
\description{
Internal function to create the summary table (bottom part).
}
