% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pareto_ok.R
\name{format_stat_value}
\alias{format_stat_value}
\title{Format Statistic Values}
\usage{
format_stat_value(x, stat_name)
}
\arguments{
\item{x}{Value to format}

\item{stat_name}{Statistic name}
}
\value{
Formatted string
}
\description{
Internal function to format statistic values for table display.
}
