% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fishbone_ok.R
\name{iQualityR.plot-fishbone}
\alias{iQualityR.plot-fishbone}
\title{Fishbone diagram functions}
\description{
Fishbone (cause-and-effect) diagram series for quality management root cause analysis.
}
\keyword{internal}
