% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hypothesis_ok.R
\name{iQualityR.plot-hypothesis}
\alias{iQualityR.plot-hypothesis}
\title{Hypothesis testing visualization functions}
\description{
Hypothesis testing visualization series including rejection region plots and boxplots with statistics tables.
}
\keyword{internal}
