% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factory.R
\name{iQualityR.plot-package}
\alias{iQualityR.plot-package}
\title{Plot Factory Functions}
\description{
A collection of modular and elegant ggplot2 plotting functions
that are consistent with the iQualityR ecosystem.
}
\keyword{internal}
