% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pareto_ok.R
\name{iQualityR.plot-pareto}
\alias{iQualityR.plot-pareto}
\title{Enhanced Pareto Chart}
\description{
Creates an enhanced Pareto chart with additional features like grouping,
weighted analysis, and interactive table display. Also supports facetting
and "Other" category aggregation.
}
\keyword{internal}
