% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_save_ok.R
\name{iQualityR.plot-save}
\alias{iQualityR.plot-save}
\title{General chart save function}
\description{
Supports DiagrammeR (grViz, mermaid, dgr_graph), ggplot2, patchwork and other objects.
}
\keyword{internal}
