% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_turtle_ok.R
\name{iQualityR.plot-turtle}
\alias{iQualityR.plot-turtle}
\title{Draw Turtle Diagram, integrated with IqrTheme theme system, supports export}
\description{
Draw Turtle Diagram, integrated with IqrTheme theme system, supports export
}
\keyword{internal}
