% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factory.R
\name{layers_boxplot}
\alias{layers_boxplot}
\title{Create Layers for Boxplot}
\usage{
layers_boxplot(
  add_jitter = TRUE,
  ...,
  boxplot_args = list(),
  jitter_args = list()
)
}
\arguments{
\item{add_jitter}{If TRUE, adds jittered points.}

\item{...}{Common arguments passed to both geoms.}

\item{boxplot_args}{List of arguments passed to geom_boxplot.}

\item{jitter_args}{List of arguments passed to geom_jitter (if used).}
}
\value{
A list of ggplot2 layers.
}
\description{
Create Layers for Boxplot
}
