% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factory.R
\name{layers_percentile_band}
\alias{layers_percentile_band}
\title{Create Layers for Percentile Band (Forest Plot)}
\usage{
layers_percentile_band(..., errorbar_args = list(), point_args = list())
}
\arguments{
\item{...}{Common arguments passed to geom_errorbar and geom_point.}

\item{errorbar_args}{List of arguments passed to geom_errorbar.}

\item{point_args}{List of arguments passed to geom_point.}
}
\value{
A list of ggplot2 layers.
}
\description{
Create Layers for Percentile Band (Forest Plot)
}
