% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factory.R
\name{layers_violin}
\alias{layers_violin}
\title{Create Layers for Violin Plot}
\usage{
layers_violin(
  add_boxplot = TRUE,
  ...,
  violin_args = list(),
  boxplot_args = list()
)
}
\arguments{
\item{add_boxplot}{If TRUE, adds a small boxplot inside.}

\item{...}{Common arguments passed to geom_violin and (optionally) geom_boxplot.}

\item{violin_args}{List of arguments passed to geom_violin.}

\item{boxplot_args}{List of arguments passed to geom_boxplot (if used).}
}
\value{
A list of ggplot2 layers.
}
\description{
Create Layers for Violin Plot
}
