% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factory.R
\name{plot_acf}
\alias{plot_acf}
\title{Create an ACF Plot}
\usage{
plot_acf(
  data_vec,
  theme = NULL,
  lag.max = NULL,
  ci = 0.95,
  highlight_sig = TRUE
)
}
\arguments{
\item{data_vec}{numeric vector}

\item{theme}{theme specification}

\item{lag.max}{maximum lag, default NULL (auto)}

\item{ci}{confidence level, default 0.95}

\item{highlight_sig}{logical, highlight bars exceeding CI}
}
\value{
ggplot object
}
\description{
Create an ACF Plot
}
