% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factory.R
\name{plot_correlation_heatmap}
\alias{plot_correlation_heatmap}
\title{Create a Correlation Heatmap
Create a Correlation Heatmap}
\usage{
plot_correlation_heatmap(data, theme = NULL, digits = 2, diagonal = FALSE)
}
\arguments{
\item{data}{data.frame}

\item{theme}{theme specification}

\item{digits}{number of digits for correlation labels}

\item{diagonal}{logical, whether to show diagonal (default FALSE, set to NA)}
}
\value{
ggplot object
}
\description{
Create a Correlation Heatmap
Create a Correlation Heatmap
}
