% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factory.R
\name{plot_interaction_line}
\alias{plot_interaction_line}
\title{Create an Interaction Plot}
\usage{
plot_interaction_line(
  data,
  x_var,
  y_var,
  group_var,
  fun = c("mean", "median"),
  theme = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{x_var}{name of x-axis variable (factor or numeric)}

\item{y_var}{name of y-axis variable (continuous)}

\item{group_var}{name of grouping variable (factor)}

\item{fun}{aggregation function, "mean" or "median"}

\item{theme}{theme specification}
}
\value{
ggplot object
}
\description{
Create an Interaction Plot
}
