% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factory.R
\name{plot_pacf}
\alias{plot_pacf}
\title{Create a PACF Plot}
\usage{
plot_pacf(
  data_vec,
  theme = NULL,
  lag.max = NULL,
  ci = 0.95,
  highlight_sig = TRUE
)
}
\arguments{
\item{data_vec}{A numeric vector.}

\item{theme}{Theme to use.}

\item{lag.max}{Max lag to calculate. If NULL, uses default.}

\item{ci}{confidence level, default 0.95}

\item{highlight_sig}{logical, highlight bars exceeding CI}
}
\value{
A ggplot object.
}
\description{
Returns a complete PACF plot object.
}
