% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factory.R
\name{plot_roc_curve}
\alias{plot_roc_curve}
\title{Create a ROC Curve Plot}
\usage{
plot_roc_curve(data, labels_var, predictions_var, theme = NULL, ci_auc = FALSE)
}
\arguments{
\item{data}{data.frame}

\item{labels_var}{name of true labels column (binary, 0/1 or factor)}

\item{predictions_var}{name of predicted probabilities column}

\item{theme}{theme specification}

\item{ci_auc}{logical, add confidence interval for AUC (bootstrap)}
}
\value{
ggplot object
}
\description{
Create a ROC Curve Plot
}
