% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pareto_ok.R
\name{process_group}
\alias{process_group}
\title{Process Group Data}
\usage{
process_group(group_data, threshold)
}
\arguments{
\item{group_data}{Data for a single group}

\item{threshold}{Threshold for "Other" aggregation}
}
\value{
Processed group data
}
\description{
Internal function to process a single group for Pareto chart.
}
