% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factory.R
\name{layers_control_chart}
\alias{layers_control_chart}
\title{Create Layers for Control Chart}
\usage{
layers_control_chart(
  data,
  ucl_color = "red",
  cl_color = "blue",
  data_color = "black"
)
}
\arguments{
\item{data}{A data.frame with columns x, y, cl, lcl, ucl.}

\item{ucl_color}{Color for UCL/LCL lines.}

\item{cl_color}{Color for center line.}

\item{data_color}{Color for data points and line.}
}
\value{
A list of ggplot2 layers.
}
\description{
Returns layers for a base control chart.
}
