% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factory.R
\name{layers_spec_limits}
\alias{layers_spec_limits}
\title{Create Layers for Specification Limits}
\usage{
layers_spec_limits(
  lsl = NULL,
  usl = NULL,
  spec_color = "red",
  lsl_label = paste("LSL =", lsl),
  usl_label = paste("USL =", usl),
  vline_args = list(),
  annotate_args = list()
)
}
\arguments{
\item{lsl}{Lower specification limit. If NULL, not drawn.}

\item{usl}{Upper specification limit. If NULL, not drawn.}

\item{spec_color}{Color for specification lines and labels.}

\item{lsl_label}{Label for LSL.}

\item{usl_label}{Label for USL.}

\item{vline_args}{List of arguments passed to geom_vline.}

\item{annotate_args}{List of arguments passed to annotate (applies to both).}
}
\value{
A list of ggplot2 layers.
}
\description{
Create Layers for Specification Limits
}
