% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factory.R
\name{layers_trend_line}
\alias{layers_trend_line}
\title{Create Layers for Trend Line}
\usage{
layers_trend_line(
  mapping = NULL,
  data = NULL,
  add_points = TRUE,
  smoothing = NULL,
  line_args = list(),
  point_args = list(),
  smooth_args = list()
)
}
\arguments{
\item{mapping}{A ggplot2 aesthetic mapping.}

\item{data}{A data frame.}

\item{add_points}{If TRUE, adds points to the line.}

\item{smoothing}{Smoothing method (e.g. "lm", "loess"). If NULL, no smoothing.}

\item{line_args}{List of arguments passed to geom_line.}

\item{point_args}{List of arguments passed to geom_point (if used).}

\item{smooth_args}{List of arguments passed to geom_smooth (if used).}
}
\value{
A list of ggplot2 layers.
}
\description{
Create Layers for Trend Line
}
