% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factory.R
\name{set_default_theme}
\alias{set_default_theme}
\title{Set Global Default Theme}
\usage{
set_default_theme(theme = "prism")
}
\arguments{
\item{theme}{theme name (character) or theme object}
}
\value{
invisible NULL
}
\description{
Set Global Default Theme
}
